requirements.txt
requests
google-cloud-storage
google-cloud-bigquery
functions-framework
google-cloud-logging
google-auth
